Instructions, Configuration and Stuff
Added 2 lines in Instructions and removed "and Tweaks/Tips" from "Configuration and Tweaks/Tips", now only "Configuration". Also added "Guide" on how to configure scheduled task in Task Scheduler, and a line to indicate that it exists in "What will this script do". Few lines also changed, which are not worthy mentioning.